endpoint bringup: Geting closer
This commit gets closer to being able to bring ports up on p4-eBPF.
We're now able to create the veth pair, put one into the "switch"
namespace, and attach  the eBPF programs to the interface in the
"switch" namespace. We can also create table entries and return the veth
to docker so the container has it attached.

Signed-off-by: Kyle Mestery <mestery@mestery.com>