plugin: Cleanup locking code
Remove a locking issue with the network map. This was done by intfCount
as a member of the nwMap struct. We no longer need this, as it was
legacy code from the vpp plugin which this is based on. Without this, I
was able to remove the nwMap lock/unlock in handlerCreateNetwork().

While here, also remove clientP4Port from the epVal struct, as we no
longer need that, and it came from the p4-dpdk version of this plugin.

Signed-off-by: Kyle Mestery <mestery@mestery.com>